Organisation of the project
