+++ Utility/Utility.csproj
@@ -78,6 +78,8 @@
     <Compile Include="Network\Function.cs" />
     <Compile Include="Network\IPacket.cs" />
     <Compile Include="Network\MessageTCP.cs" />
+    <Compile Include="Network\Server\ServerComunication.cs" />
+    <Compile Include="Network\Server\ServerSocket.cs" />
     <Compile Include="Network\Token.cs" />
     <Compile Include="Network\Users\Rank.cs" />
     <Compile Include="Network\Users\User.cs" />
